Refactor: Extract Precise class to separate header
- Move Precise class and related types to npsr/precise.h
- Add deduction guides for cleaner template instantiation
- Improve modularity by separating precision control from common utilities
